feat(compliance): FP-7360 queries (#9054)
## Describe your changes
- This PR adds some other queries to the `app/partners` application.
- It also renames the existent ones due to this PR **#9010** 👈 from
backend

+++ frontend/apps/partners/src/queries/delete-partner-doc-templates.ts
@@ -1,3 +1,4 @@
+import { getAuthCookie } from '@/app/actions';
 import { DASHBOARD_AUTHORIZATION_HEADER } from '@/config/constants';
 
 import baseFetch from './base-fetch';
@@ -5,21 +6,21 @@ import baseFetch from './base-fetch';
 type EmptyResponse = Record<string, never>;
 
 /**
- * Deletes compliance document templates.
+ * Deactivates a compliance document template.
  *
- * @param {string} authToken - the authentication token
  * @param {string} templateId - the ID of the template to delete
  * @return {Promise<EmptyResponse>} a Promise that resolves with an empty response
  */
-const deleteComplianceDocTemplates = async (
-  authToken: string,
-  templateId: string,
-) =>
-  authToken && templateId
+const deletePartnerDocTemplates = async (templateId: string) => {
+  const token = await getAuthCookie();
+  if (!token) return Promise.reject(new TypeError('Missing auth token'));
+
+  return templateId
     ? baseFetch<EmptyResponse>(`/partner/doc_templates/${templateId}`, {
-        headers: { [DASHBOARD_AUTHORIZATION_HEADER]: authToken },
+        headers: { [DASHBOARD_AUTHORIZATION_HEADER]: token },
         method: 'DELETE',
       })
     : Promise.reject(new TypeError('Missing required parameters'));
+};
 
-export default deleteComplianceDocTemplates;
+export default deletePartnerDocTemplates;
