+++ frontend/apps/partners/src/queries/delete-partner-partnerships-requests.ts
@@ -1,3 +1,4 @@
+import { getAuthCookie } from '@/app/actions';
 import { DASHBOARD_AUTHORIZATION_HEADER } from '@/config/constants';
 
 import baseFetch from './base-fetch';
@@ -5,26 +6,28 @@ import baseFetch from './base-fetch';
 type EmptyResponse = Record<string, never>;
 
 /**
- * Deletes compliance partner requests.
+ * Retracts a document request.
  *
- * @param {string} authToken - the authentication token
  * @param {string} partnershipId - the partnership ID
  * @param {string} requestId - the request ID
  * @return {Promise<EmptyResponse>} a promise that resolves to an empty response or rejects with a TypeError
  */
-const deleteCompliancePartnersRequests = async (
-  authToken: string,
+const deletePartnerPartnershipsRequests = async (
   partnershipId: string,
   requestId: string,
-) =>
-  authToken && partnershipId && requestId
+) => {
+  const token = await getAuthCookie();
+  if (!token) return Promise.reject(new TypeError('Missing auth token'));
+
+  return partnershipId && requestId
     ? baseFetch<EmptyResponse>(
         `/partner/partnerships/${partnershipId}/requests/${requestId}`,
         {
-          headers: { [DASHBOARD_AUTHORIZATION_HEADER]: authToken },
+          headers: { [DASHBOARD_AUTHORIZATION_HEADER]: token },
           method: 'DELETE',
         },
       )
     : Promise.reject(new TypeError('Missing required parameters'));
+};
 
-export default deleteCompliancePartnersRequests;
+export default deletePartnerPartnershipsRequests;
